add dispatcher on compute node
